refactor(frontend): forms/search を .ts 化 (Phase 3-B)
- showSearchPicker() の戻り値を Promise<SearchPreset | null> に型付け
- collectSearch は CollectedConfig.search に SearchEntry を投入
- HTMLInputElement / HTMLSelectElement / HTMLUListElement に narrowing
